requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain and Related
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
langchain-pinecone>=0.0.1
langchain-groq>=0.0.1

# Vector Database
pinecone-client>=3.0.0

# Document Processing
pypdf>=3.17.0

# Embeddings
sentence-transformers>=2.2.2
transformers>=4.35.0

# Additional Dependencies
huggingface-hub>=0.19.0
